Rename theme-orange-derived in theme-orange-country
