+++ settings.gradle.kts
@@ -37,5 +37,5 @@ include(":foundation")
 include(":core")
 include(":theme-contract")
 include(":theme-orange")
-include(":theme-orange-derived")
+include(":theme-orange-country")
 include(":theme-white-label")
