+++ theme-orange-country/src/main/AndroidManifest.xml
